Qt 5.6.2 Release

+++ src/gui/painting/qpainter_p.h
@@ -53,7 +53,6 @@
 #include "QtGui/qpainter.h"
 #include "QtGui/qpainterpath.h"
 #include "QtGui/qpaintengine.h"
-#include <QtCore/qhash.h>
 
 #include <private/qpen_p.h>
 
@@ -241,7 +240,7 @@ public:
     }
 
     QTransform viewTransform() const;
-    int effectiveDevicePixelRatio() const;
+    qreal effectiveDevicePixelRatio() const;
     QTransform hidpiScaleTransform() const;
     static bool attachPainterPrivate(QPainter *q, QPaintDevice *pdev);
     void detachPainterPrivate(QPainter *q);
